refactor(server): move tag manage to admin #90
