--- .bruno/Feed/Tag/Delete Tag.bru
@@ -1,15 +1,0 @@
-meta {
-  name: Delete Tag
-  type: http
-  seq: 1
-}
-
-delete {
-  url: http://localhost:{{PORT}}/api/tags/:code
-  body: none
-  auth: inherit
-}
-
-params:path {
-  code: faith
-}
