+++ .bruno/Feed/Tag/Read All Tag.bru
@@ -1,7 +1,7 @@
 meta {
   name: Read All Tag
   type: http
-  seq: 6
+  seq: 2
 }
 
 get {
