+++ .bruno/Feed/Tag/Read Tag.bru
@@ -1,7 +1,7 @@
 meta {
   name: Read Tag
   type: http
-  seq: 5
+  seq: 1
 }
 
 get {
