+++ src/main/java/app/handong/feed/controller/TbadminController.java
@@ -1,10 +1,13 @@
 package app.handong.feed.controller;
 
+import app.handong.feed.dto.TagDto;
 import app.handong.feed.dto.TbadminDto;
 import app.handong.feed.security.annotation.RequiredUserScopes;
 import app.handong.feed.security.enums.UserScope;
+import app.handong.feed.service.TagService;
 import app.handong.feed.service.TbadminService;
 import app.handong.feed.util.RequestUtils;
+import io.swagger.v3.oas.annotations.Operation;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -17,10 +20,9 @@ import java.util.Map;
 public class TbadminController {
     private final TbadminService tbadminService;
 
-    public TbadminController(TbadminService tbadminService) {
+    public TbadminController(TbadminService tbadminService, TagService tagService) {
         this.tbadminService = tbadminService;
     }
-
 
     @GetMapping("/users")
     @RequiredUserScopes({
@@ -75,5 +77,37 @@ public class TbadminController {
         String reqUserId = RequestUtils.getReqUserId(request);
         tbadminService.deleteApiKey(reqUserId, id);
         return ResponseEntity.noContent().build();
+    }
+
+    @PostMapping("/tags")
+    @Operation(summary = "태그 생성")
+    public ResponseEntity<TagDto.CreateResDto> createTag(@RequestBody TagDto.CreateReqDto dto, HttpServletRequest request) {
+        String reqUserId = RequestUtils.getReqUserId(request);
+        return ResponseEntity.ok(tbadminService.createTag(dto));
+    }
+
+    @PostMapping("/tags/batch")
+    @Operation(
+            summary = "태그 여러 개 생성",
+            description = "여러 개의 태그를 한 번에 생성합니다. 각 태그는 고유한 code 값을 가져야 합니다."
+    )
+    public ResponseEntity<List<TagDto.CreateResDto>> createTags(@RequestBody List<TagDto.CreateReqDto> requestList, HttpServletRequest request) {
+        String reqUserId = RequestUtils.getReqUserId(request);
+        return ResponseEntity.ok(tbadminService.createTags(requestList));
+    }
+
+    @PatchMapping("/tags/{code}")
+    @Operation(summary = "태그 수정")
+    public ResponseEntity<TagDto.UpdateResDto> updateTag(@PathVariable String code,
+                                                         @RequestBody TagDto.UpdateReqDto dto, HttpServletRequest request) {
+        String reqUserId = RequestUtils.getReqUserId(request);
+        return ResponseEntity.ok(tbadminService.updateTag(code, dto));
+    }
+
+    @DeleteMapping("/tags/{code}")
+    @Operation(summary = "태그 삭제")
+    public ResponseEntity<TagDto.DeleteResDto> deleteTag(@PathVariable String code, HttpServletRequest request) {
+        String reqUserId = RequestUtils.getReqUserId(request);
+        return ResponseEntity.ok(tbadminService.deleteTag(code));
     }
 }
