+++ src/main/java/app/handong/feed/service/TagService.java
@@ -7,10 +7,6 @@ import java.util.List;
 
 @Service
 public interface TagService {
-    TagDto.CreateResDto createTag(TagDto.CreateReqDto dto);
-    List<TagDto.CreateResDto> createTags(List<TagDto.CreateReqDto> requestList);
     TagDto.ReadResDto readTag(String code);
     List<TagDto.ReadResDto> readAllTags();
-    TagDto.UpdateResDto updateTag(String code, TagDto.UpdateReqDto dto);
-    TagDto.DeleteResDto deleteTag(String code);
 }
