+++ src/main/java/app/handong/feed/service/TbadminService.java
@@ -1,5 +1,6 @@
 package app.handong.feed.service;
 
+import app.handong.feed.dto.TagDto;
 import app.handong.feed.dto.TbadminDto;
 import org.springframework.stereotype.Service;
 
@@ -16,4 +17,10 @@ public interface TbadminService {
     TbadminDto.ApiKeyCreateRespDto issueApiKey(String userId, TbadminDto.ApiKeyCreateReqDto req);
     TbadminDto.ApiKeyDetail toggleApiKeyStatus(String userId, String apiKeyId);
     void deleteApiKey(String userId, String apiKeyId);
+
+    TagDto.CreateResDto createTag(TagDto.CreateReqDto dto);
+    List<TagDto.CreateResDto> createTags(List<TagDto.CreateReqDto> requestList);
+    TagDto.UpdateResDto updateTag(String code, TagDto.UpdateReqDto dto);
+    TagDto.DeleteResDto deleteTag(String code);
+
 }
