+++ src/main/java/app/handong/feed/service/impl/TagServiceImpl.java
@@ -2,7 +2,6 @@ package app.handong.feed.service.impl;
 
 import app.handong.feed.domain.Tag;
 import app.handong.feed.dto.TagDto;
-import app.handong.feed.exception.data.DuplicateTagCodeException;
 import app.handong.feed.exception.data.NotFoundException;
 import app.handong.feed.repository.TagRepository;
 import app.handong.feed.service.TagService;
@@ -18,33 +17,6 @@ import java.util.stream.Collectors;
 public class TagServiceImpl implements TagService {
 
     private final TagRepository tagRepository;
-
-    @Override
-    public TagDto.CreateResDto createTag(TagDto.CreateReqDto dto) {
-        if (tagRepository.existsById(dto.getCode())) {
-            throw new DuplicateTagCodeException(dto.getCode());
-        }
-
-        return TagDto.CreateResDto.fromEntity(tagRepository.save(dto.toEntity()));
-    }
-
-    public List<TagDto.CreateResDto> createTags(List<TagDto.CreateReqDto> requestList) {
-        List<Tag> tags = requestList.stream()
-                .map(TagDto.CreateReqDto::toEntity)
-                .toList();
-
-        for (Tag tag : tags) {
-            if (tagRepository.existsById(tag.getCode())) {
-                throw new DuplicateTagCodeException(tag.getCode());
-            }
-        }
-
-        List<Tag> savedTags = tagRepository.saveAll(tags);
-        return savedTags.stream()
-                .map(TagDto.CreateResDto::fromEntity)
-                .toList();
-    }
-
 
     @Override
     @Transactional(readOnly = true)
@@ -79,38 +51,5 @@ public class TagServiceImpl implements TagService {
                         tag.getUpdatedAt()
                 ))
                 .collect(Collectors.toList());
-    }
-
-    @Override
-    @Transactional
-    public TagDto.UpdateResDto updateTag(String code, TagDto.UpdateReqDto dto) {
-        Tag tag = tagRepository.findById(code)
-                .orElseThrow(() -> new NotFoundException("Tag not found with code: " + code));
-
-        // 필드 업데이트
-        tag.setLabel(dto.getLabel());
-        tag.setUserDesc(dto.getUserDesc());
-        tag.setLlmDesc(dto.getLlmDesc());
-        tag.setColorHex(dto.getColorHex());
-        tag.setPriorityWeight(dto.getPriorityWeight());
-
-        return new TagDto.UpdateResDto(
-                tag.getCode(),
-                tag.getLabel(),
-                tag.getUserDesc(),
-                tag.getLlmDesc(),
-                tag.getColorHex(),
-                tag.getPriorityWeight(),
-                tag.getUpdatedAt()
-        );
-    }
-
-    @Override
-    public TagDto.DeleteResDto deleteTag(String code) {
-        tagRepository.findById(code)
-            .orElseThrow(() -> new NotFoundException("Tag not found with code: " + code));
-
-        tagRepository.deleteById(code);
-        return new TagDto.DeleteResDto(code, "Tag deleted successfully");
     }
 }
